Add log_versions
