+++ rex/utilities/utilities.py
@@ -8,6 +8,7 @@ import json
 import os
 import numpy as np
 import pandas as pd
+from pandas.api.types import CategoricalDtype
 import re
 from scipy.spatial import cKDTree
 import time
@@ -766,3 +767,72 @@ def res_dist_threshold(lat_lons, tree=None, margin=1.05):
     dists = dists[(dists != 0)]
 
     return margin * (2 ** 0.5) * (dists.max() / 2)
+
+
+def get_dtype(col):
+    """
+    Get column dtype for converstion to records array
+
+    Parameters
+    ----------
+    col : pandas.Series
+        Column from pandas DataFrame
+
+    Returns
+    -------
+    out : str
+        String representation of converted dtype for column:
+        -  float = float32
+        -  int = int16 or int32 depending on data range
+        -  object/str = U* max length of strings in col
+    """
+    dtype = col.dtype
+
+    if isinstance(dtype, CategoricalDtype):
+        col = col.astype(type(col.values[0]))
+        out = get_dtype(col)
+    elif np.issubdtype(dtype, np.floating):
+        out = 'float32'
+    elif np.issubdtype(dtype, np.integer):
+        if col.max() < 32767:
+            out = 'int16'
+        else:
+            out = 'int32'
+    elif np.issubdtype(dtype, np.object_):
+        size = int(col.astype(str).str.len().max())
+        out = 'S{:}'.format(size)
+    else:
+        out = dtype
+
+    return out
+
+
+def to_records_array(df):
+    """
+    Convert pandas DataFrame to numpy Records Array
+
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        Pandas DataFrame to be converted
+
+    Returns
+    -------
+    numpy.rec.array
+        Records array of input df
+    """
+    meta_arrays = []
+    dtypes = []
+    for c_name, c_data in df.iteritems():
+        dtype = get_dtype(c_data)
+
+        if np.issubdtype(dtype, np.bytes_):
+            data = c_data.astype(str).str.encode('utf-8').values
+        else:
+            data = c_data.values
+
+        arr = np.array(data, dtype=dtype)
+        meta_arrays.append(arr)
+        dtypes.append((c_name, dtype))
+
+    return np.core.records.fromarrays(meta_arrays, dtype=dtypes)
